requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Database (for future use)
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Machine Learning and NLP
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0

# Text processing
nltk==3.8.1
spacy==3.7.2

# Time series analysis
prophet==1.1.4
neuralprophet==0.6.0

# Topic modeling
bertopic==0.15.0
gensim==4.3.2

# LLM and RAG
openai==1.3.7
langchain==0.0.350
chromadb==0.4.18
faiss-cpu==1.7.4

# Reinforcement Learning
ray[rllib]==2.8.0

# MLOps
mlflow==2.8.1
dvc==3.27.0

# Monitoring and logging
prometheus-client==0.19.0
grafana-api==1.0.3

# Privacy and security
opacus==1.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
requests==2.31.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Docker and deployment
gunicorn==21.2.0
